Insert ad slots straight after article fetched
